More Lenient Version Resolution
Previously, version resolution with the contents of repository indexes was
strict.  If there was an illegal version, all resolution would stop; in other
words, the buildpack would cease to function.  This recently became an issue
for the first time when a third-party repository added an invalid version
number to their index breaking all users of the buildpack.  This change
updates the version resolution logic to ignore and warn on a bad version
instead of failing.

[#290]