Include More Classes In Calculation
Previously, only the classes actually in the application were candidates for
use when calculating the total number of classes in the application.  There
was a fixed 5500 class offset to account for the un-evaluated JRE.  However,
there were problems when downstream buildpacks added large servers to the
application (e.g. Wildfly) and 10s of thousands classes were not accounted
for.  This change enumerates the number of classes in the _entire_ droplet,
including the JRE, and removes the constant offset.