Print JVM Memory Configuration on Startup
Previously, the JVM memory configuration was printed on startup by the memory
calculator itself.  Because the memory calculator stdout is captured, this
printing had to be done to stderr.  This caused some alarm (mostly by Sabby),
so this alternative was created.  This change prints the captured output (plus
some other verbage), removing the need for the memory calculator to do it.