Container Security Provider
This change adds a Container Security Provider to the buildpack.  This
security provider is aware of container identity in a Cloud Foundry container
as defined by a private key and certificate.  As part of the addition of this
functionality, extensions were made to the Droplet Public API to allow
components to contribute both extension directories and security providers.
This resulted in the update of the Dyadic, Luna, and ProtectApp providers.