Configure Class Count
Previously, the class_count memory calculator parameter was not defined in the
configuration YAML files where it was legal. This meant that the buildpack
automatically discarded any attempt to configure it, as it was not considered
a legal value.  This change adds the key to the configuration, with an empty
value ensure that is configurable with an environment variable and defaults to
the calculated value when it is not.

[resolves #464]