Disable JVM DNS Caching
Previously JVM DNS caching was always governed by the JVM's settings of
always/10s.  This causes issues in situations where DNS entries are being
refreshed by BOSH quite quickly.  To address this, BOSH has added a DNS server
of its own that is accessible on a link-local address.  This BOSH DNS server
implements the appropriate caching close by (and obeying ttls) removing the
need for the JVM to so.  This change notices the BOSH DNS as staging time and
if it exists, disables the default JVM DNS caching in lieu of the BOSH DNS
caching.

[resolves #505]