Retry with Exponential Backoff
Previously, when a download failed for a network reason, it would be retried
up to 5 times, but they would happen as quickly as possible after the failure.
Given that network errors are transient but typically take a perceptible
amount of time this retry almost always experienced the same error every
single time.  This change adds an exponential backoff to the retry behavior
waiting between 5 and 60 seconds as more retried are attempted.

[resolves #524]