Explicitly Set MALLOC_ARENA_MAX
Because the JVM isn't very aware of the memory model of the OS that it's
running on, it would often request memory in a way that caused a large number
of malloc arenas, especially under a full heap.  Unfortunately, it would not
use the malloc arena count in it's internal accounting of it's current memory
usage, causing the container to kill the process.  This change fixes the
number of arenas, taking the performance hit for better behavior at the edge
of a container.

[resolves #550]