Support for TCPKeepAlive configuration
This enhancement supports configuring the TCPKeepAlive setting in Luna
Security Provider.  Configuration is done by setting the
tcp_keep_alive_enabled configuration parameter to true of false (default is
false).  Default behavior sets the TCPKeepAlive Luna client parameter to 0
which is the default setting if the parameter is not present in the
Chryptoki.conf file.

[#589]