Security Provider Classpath
Previously, an attempt to ensure that the security providers showed in the new
Java 9+ classpath (not in an ext directory) did not work.  Instead, they were
randomly spread among various classpaths depending on which container was
being used.  This change introduces a root_libraries abstraction that
identifies dependencies that MUST be in the SYSTEM classloader, and then
updates all of the containers to ensure that they are.

[resolves #585]