Tomcat External Configuration Offline
Previously, Tomcat External Configuration was subject to the
InternetAvailablity flag of the buildpack.  This meant that when using an
offline buildpack (as most do), external configuration could never be
downloaded.  Since external configuration would almost exclusively come from
within the firewall, this enforcement is overkill.  This change updates the
implementation to download regardless of the InternetAvailablity.

[resolves #638]

Signed-off-by: Ben Hale <bhale@pivotal.io>