Multi-buildpack Paths
Finally found it!  Turns out that when the multi-buildpack added paths from
previous supply buildpacks, it added them as raw Pathnames and not
FilteringPathnames.  Once this was done, the order of addition and the sorting
algorithm would cause intermittent failures because FilteringPathnames could
sort with Pathnames, but not vice versa.  This change ensure that the
Multi-buildpack Framework adds FilteringPathnames to all collections.

Signed-off-by: Ben Hale <bhale@pivotal.io>