Restore unique_host_name logic (#884)
PR #870 reversed the logic overriding the default_unique_host_name with
the application_name.  This results in problems with AppDynamics where
applications running on multiple foundations are unable to add metrics
when using the same application name in Cloud Foundry.

Really, application_name should never be used. unique_host_name is
by definition a unique identifier vs an application_name, which is not
guaranteed to be unique.

But this restores the previous logic, which was functional and may have
handled cases I am not aware of.