Version matcher now prefers a specific match over a partial match (#907)
Previous to this commit, if you have two versions say `3.1.1` and `3.1.1_BETA` the version matcher would incorrectly select `3.1.1_BETA`.  It would behave as if a wildcard was present even though it's not present.

This commit ensures that the same scenario will select version `3.1.1`, which is an exact match for the version requested.

This may result in a difference of behavior for some users. If you are impacted by this, you need to switch your requested version to `3.1.1_+`. Using the scenario above, this will return `3.1.1_BETA`.

Signed-off-by: Daniel Mikusa <dmikusa@vmware.com>